Add card link to localstorage once the quiz is graded, if passed
