+++ src/pages/Main/Main.tsx
@@ -1,6 +1,6 @@
 import "./Main.scss";
 
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 
 import { CardItemData, CardList } from "../../components/Card";
 import { MainHeader } from "../../components/MainHeader";
@@ -13,7 +13,6 @@ function Main() {
     activeQuiz: CardItemData | null;
     showModal: boolean;
   }>({ activeQuiz: null, showModal: false });
-  const [lastCompletedCardLink, setLastCompletedCardLink] = useState("");
 
   const handleSelect = (card: CardItemData) => {
     setQuizState({ activeQuiz: card, showModal: true });
@@ -23,31 +22,6 @@ function Main() {
     // TODO: add this to the user's history
     logger.info(`Removing ${card.title}`);
   };
-
-  const getCardLinks = (): string[] => {
-    return JSON.parse(localStorage.getItem("eras.completedCardLinks") || "[]");
-  };
-
-  const storeCardLinks = (cardLinks: string[]) => {
-    localStorage.setItem("eras.completedCardLinks", JSON.stringify(cardLinks));
-  };
-
-  useEffect(() => {
-    const completedCardLinks = getCardLinks();
-
-    completedCardLinks.push(lastCompletedCardLink);
-    storeCardLinks(completedCardLinks);
-
-    return () => {
-      const cardLinksToCleanUp = getCardLinks();
-
-      const cleanedUpCardLinks = cardLinksToCleanUp.filter((link) => {
-        return link !== "";
-      });
-
-      storeCardLinks(cleanedUpCardLinks);
-    };
-  }, [lastCompletedCardLink]);
 
   return (
     <div className="main" role="main">
@@ -62,9 +36,7 @@ function Main() {
         <QuizModal
           card={quizState.activeQuiz}
           show={quizState.showModal}
-          handleClose={(isCompleted: boolean) => {
-            const link = quizState.activeQuiz?.link;
-            if (isCompleted && link) setLastCompletedCardLink(link);
+          handleClose={() => {
             setQuizState({ activeQuiz: null, showModal: false });
           }}
         />
